backend/requirements.txt
pandas==2.2.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
schedule==1.2.1
selenium==4.18.1
fastapi==0.104.1
uvicorn==0.24.0